CustomScore sql优化

+++ src/main/java/org/example/java2final/vo/CustomScoreVO.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @Data
 public class CustomScoreVO {
     private Integer accountAgeDays;
-    private Double qualityMetric;
+    private Float qualityMetric;
 }
